crypto_factor_strategy/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
arch==7.2.0
astropy==7.0.1
astropy-iers-data==0.2025.4.21.0.37.6
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blosc2==3.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cloudpickle==3.1.1
coinbase==2.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cvxpy==1.6.5
cycler==0.12.1
dask==2025.4.0
dateparser==1.2.1
decorator==5.2.1
dill==0.4.0
distro==1.9.0
duckdb==1.2.2
ecos==2.0.14
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
future==1.0.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
humanize==4.12.2
hyperopt==0.2.7
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
llvmlite==0.44.0
lmdb==1.6.2
lmdbm==0.0.6
locket==1.0.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multiprocess==0.70.18
multitasking==0.0.11
mypy_extensions==1.1.0
ndindex==1.9.2
networkx==3.4.2
numba==0.61.2
numexpr==2.10.2
numpy==2.2.5
openai==1.75.0
openpyxl==3.1.5
optuna==4.3.0
osqp==1.0.3
packaging==25.0
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pathos==0.3.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyarrow==19.0.1
pybind11==2.13.6
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.3
pyportfolioopt==1.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
QuantStats==0.0.64
regex==2024.11.6
requests==2.32.3
Riskfolio-Lib==7.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
scs==3.2.7.post2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
ta-lib==0.6.3
tables==3.10.2
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vectorbtpro @ git+https://github.com/polakowo/vectorbt.pro.git@6e344a8230eaf718593f4570378486ee1d4178f6
wcwidth==0.2.13
websocket-client==1.8.0
widgetsnbextension==4.0.14
XlsxWriter==3.2.3
yfinance==0.2.55
zipp==3.21.0
